my-agents/requirements.txt
google-adk
   toolbox-core
fastapi
uvicorn
google-cloud-aiplatform
google-genai
python-dotenv